move loginTest back to api key
