+++ src/test/scala/com/cognite/sdk/scala/v1/ApiKeysTest.scala
@@ -3,16 +3,8 @@
 
 package com.cognite.sdk.scala.v1
 
-import cats.Id
-import com.cognite.sdk.scala.common._
+import com.cognite.sdk.scala.common.{ReadBehaviours, ApiKeysTestSpec}
 
-class ApiKeysTest extends SdkTestSpec with ReadBehaviours {
-  private val apiKey = Option(System.getenv("TEST_API_KEY"))
-                  .getOrElse(throw new RuntimeException("TEST_API_KEY not set"))
-  override lazy val auth: Auth = ApiKeyAuth(apiKey)
-
-  override lazy val client = GenericClient.forAuth[Id](
-    "scala-sdk-test", auth)
-
+class ApiKeysTest extends ApiKeysTestSpec with ReadBehaviours {
   "ApiKeys" should behave like readable(client.apiKeys, supportsLimit = false)
 }
